+++ tests/cogs/announce_day/announce_day_test.py
@@ -8,14 +8,6 @@ from tests.duckmock.datetime import patch_now
 
 
 @pytest.fixture
-def setup_general_channel(bot, guild, text_channel):
-    bot.get_all_channels.return_value = [text_channel]
-    guild.name = "Friends Chat"
-    text_channel.guild = guild
-    text_channel.name = "general"
-
-
-@pytest.fixture
 @mock.patch("duckbot.cogs.dogs.DogPhotos")
 def dog_photos(d):
     return d
@@ -38,7 +30,7 @@ async def test_cog_unload_cancels_task(bot, dog_photos):
 @pytest.mark.asyncio
 @mock.patch("random.random", return_value=0.5)
 @mock.patch("random.choice", side_effect=["today", "tomorrow", "yesterday", "{0} {1} {2}"])
-async def test_on_hour_7am_eastern_special_day(random, choice, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_special_day(random, choice, bot, dog_photos, text_channel, setup_expected_general_channel):
     with patch_now(datetime.datetime(2002, 1, 1, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
         await clazz.on_hour()
@@ -48,7 +40,7 @@ async def test_on_hour_7am_eastern_special_day(random, choice, bot, dog_photos, 
 @pytest.mark.asyncio
 @mock.patch("random.random", return_value=0.5)
 @mock.patch("random.choice", side_effect=["today", "tomorrow", "yesterday", "{0} {1} {2}"])
-async def test_on_hour_7am_eastern_not_special_day(random, choice, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_not_special_day(random, choice, bot, dog_photos, text_channel, setup_expected_general_channel):
     with patch_now(datetime.datetime(2002, 1, 21, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
         await clazz.on_hour()
@@ -57,7 +49,7 @@ async def test_on_hour_7am_eastern_not_special_day(random, choice, bot, dog_phot
 
 @pytest.mark.asyncio
 @mock.patch("random.random", side_effect=[0.09, 0.1])
-async def test_on_hour_7am_eastern_send_dog_photo(random, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_send_dog_photo(random, bot, dog_photos, text_channel, setup_expected_general_channel):
     with patch_now(datetime.datetime(2002, 1, 21, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
         await clazz.on_hour()
@@ -67,7 +59,7 @@ async def test_on_hour_7am_eastern_send_dog_photo(random, bot, dog_photos, text_
 
 @pytest.mark.asyncio
 @mock.patch("random.random", side_effect=[0.09, 0.1])
-async def test_on_hour_7am_eastern_send_dog_photo_failure(random, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_send_dog_photo_failure(random, bot, dog_photos, text_channel, setup_expected_general_channel):
     dog_photos.get_dog_image.side_effect = RuntimeError("ded")
     with patch_now(datetime.datetime(2002, 1, 21, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
@@ -78,7 +70,7 @@ async def test_on_hour_7am_eastern_send_dog_photo_failure(random, bot, dog_photo
 
 @pytest.mark.asyncio
 @mock.patch("random.random", side_effect=[0.1, 0.09])
-async def test_on_hour_7am_eastern_send_gif(random, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_send_gif(random, bot, dog_photos, text_channel, setup_expected_general_channel):
     with patch_now(datetime.datetime(2002, 1, 25, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
         await clazz.on_hour()
@@ -89,7 +81,7 @@ async def test_on_hour_7am_eastern_send_gif(random, bot, dog_photos, text_channe
 @pytest.mark.asyncio
 @mock.patch("duckbot.cogs.announce_day.announce_day.days")
 @mock.patch("random.random", side_effect=[0.1, 0.09])
-async def test_on_hour_7am_eastern_no_gifs_to_send(random, days, bot, dog_photos, text_channel, setup_general_channel):
+async def test_on_hour_7am_eastern_no_gifs_to_send(random, days, bot, dog_photos, text_channel, setup_expected_general_channel):
     days.__getitem__.return_value = {"names": ["name"], "templates": [], "gifs": []}
     with patch_now(datetime.datetime(2002, 1, 25, hour=7)):
         clazz = AnnounceDay(bot, dog_photos)
