+++ tests/fixtures/async_mock.py
@@ -1,0 +1,15 @@
+from unittest import mock
+
+import pytest
+
+
+@pytest.fixture(scope="session", autouse=True)
+def async_mock_await_fix():
+    """Make it so @mock.patch works for async methods."""
+
+    async def async_magic():
+        """This method's __await__ is used to hack mock awaits. No implementation needed."""
+        pass
+
+    mock.MagicMock.__await__ = lambda x: async_magic().__await__()
+    yield
