+++ tests/fixtures/discord/__init__.py
@@ -1,0 +1,17 @@
+from .author import member, user
+from .bot import bot, bot_spy
+from .channel import (
+    channel,
+    dm_channel,
+    group_channel,
+    skip_if_private_channel,
+    text_channel,
+    thread,
+)
+from .context import context
+from .embed import patch_embed_equals
+from .emoji import emoji
+from .general_channel import setup_expected_general_channel
+from .guild import guild
+from .message import message
+from .voice import voice_channel, voice_client
